Añadidos los bordes y mejorada la función random para tamaños superiores a 250x250

+++ common.hpp
@@ -40,7 +40,7 @@
 
 extern const double FPS;
 
-typedef unsigned short int my_size_t;
+typedef unsigned int my_size_t;
 
 extern bool undefined;
